Track upstream pygit2 v0.20.3 release
Fix dumb packaging problems

grab some stuff out of the gitignores for the submodules to prune in the manifest so we don't get cached compiled crap in the python source distribution.

as a bonus, pruing a bunch of stuff we don't need means we're under 1MB now.

oh yeah and we don't build libgit2's tests, so installing is way faster. woo hoo!